renamed ClosedQAnsweredCard -> AnswerShared; working
